+++ kusanagi/core/payload/cmd/loader.py
@@ -68,7 +68,7 @@ def _load(placeholders: Dict[str, str]) -> List[DsPayload]:
     return payloads
 
 
-def _get_payload_from_item(item: Dict[str, List]) -> DsPayload:
+def _get_payload_from_item(item: Dict[str, Any]) -> DsPayload:
     """Returns DsPayload data structure from item."""
     filters = DsPayloadFilter(
         item["filters"]["cmd"],
